+++ Core/From.php
@@ -3,7 +3,39 @@ declare(strict_types = 1);
 
 namespace Dasuos\Sql;
 
-interface From extends Clause {
+final class From implements Statement {
 
-	public function join(Join ...$join): From;
+	private const NO_CLAUSE = '';
+
+	private $sql;
+
+	public function __construct(string $table) {
+		$this->sql['table'] = $table;
+	}
+
+	public function join(Join ...$joins): From {
+		$this->sql['joins'] = implode(
+			' ',
+			array_map(
+				static function(Join $join) {
+					return $join->sql();
+				},
+				$joins
+			)
+		);
+		return $this;
+	}
+
+	public function sql(): string {
+		return sprintf(
+			'FROM %s',
+			trim(
+				sprintf(
+					'%s %s',
+					$this->sql['table'],
+					$this->sql['joins'] ?? self::NO_CLAUSE
+				)
+			)
+		);
+	}
 }
