+++ Core/OrderBy.php
@@ -3,8 +3,31 @@ declare(strict_types = 1);
 
 namespace Dasuos\Sql;
 
-interface OrderBy extends Clause {
+final class OrderBy implements Clause {
 
-	public function ascending(): OrderBy;
-	public function descending(): OrderBy;
+	private const DEFAULT = 'ASC';
+
+	private $sql;
+
+	public function __construct(string ...$columns) {
+		$this->sql['columns'] = $columns;
+	}
+
+	public function ascending(): OrderBy {
+		$this->sql['sort'] = self::DEFAULT;
+		return $this;
+	}
+
+	public function descending(): OrderBy {
+		$this->sql['sort'] = 'DESC';
+		return $this;
+	}
+
+	public function sql(): string {
+		return sprintf(
+			'ORDER BY %s %s',
+			implode(', ', $this->sql['columns']),
+			$this->sql['sort'] ?? self::DEFAULT
+		);
+	}
 }
