+++ Tests/Unit/InsertParameters.phpt
@@ -11,12 +11,12 @@ require __DIR__ . '/../bootstrap.php';
 /**
  * @testCase
  */
-final class FlatParameters extends Tester\TestCase {
+final class InsertParameters extends Tester\TestCase {
 
 	public function testFlatteningNestedTwoDimensionalArray(): void {
 		Tester\Assert::same(
 			[0, 1, 2, 3, 4, 5, 6, 7],
-			(new Sql\FlatParameters(
+			(new Sql\InsertParameters(
 				[0, ['one' => 1, 'two' => 2], 3, 4, [5, 6], 7]
 			))->values()
 		);
@@ -25,7 +25,7 @@ final class FlatParameters extends Tester\TestCase {
 	public function testFlatteningNestedThreeAndMoreDimensionalArray(): void {
 		Tester\Assert::same(
 			[0, 1, 2, 3, 4, 5, 6, 'seven'],
-			(new Sql\FlatParameters(
+			(new Sql\InsertParameters(
 				[0, [1, [2]], [3, 4, [5, [6]]], 'seven']
 			))->values()
 		);
@@ -34,11 +34,11 @@ final class FlatParameters extends Tester\TestCase {
 	public function testFlatteningAlreadyFlatArray(): void {
 		Tester\Assert::same(
 			[0, 1, 2, 3, 4, 5, 'six', 7],
-			(new Sql\FlatParameters(
+			(new Sql\InsertParameters(
 				[0, 1, 2, 3, 4, 5, 'six', 7]
 			))->values()
 		);
 	}
 }
 
-(new FlatParameters)->run();
+(new InsertParameters)->run();
