+++ Tests/Unit/Join.phpt
@@ -11,19 +11,19 @@ require __DIR__ . '/../bootstrap.php';
 /**
  * @testCase
  */
-final class BuiltJoin extends Tester\TestCase {
+final class Join extends Tester\TestCase {
 
 	public function testReturningJoin(): void {
 		Tester\Assert::same(
 			'JOIN foobar ON foo = bar',
-			(new Sql\BuiltJoin('foobar', 'foo', 'bar'))->sql()
+			(new Sql\Join('foobar', 'foo', 'bar'))->sql()
 		);
 	}
 
 	public function testReturningLeftJoin(): void {
 		Tester\Assert::same(
 			'LEFT JOIN foobar ON foo = bar',
-			(new Sql\BuiltJoin('foobar', 'foo', 'bar'))
+			(new Sql\Join('foobar', 'foo', 'bar'))
 				->left()
 				->sql()
 		);
@@ -32,11 +32,11 @@ final class BuiltJoin extends Tester\TestCase {
 	public function testReturningRightJoin(): void {
 		Tester\Assert::same(
 			'RIGHT JOIN foobar ON foo = bar',
-			(new Sql\BuiltJoin('foobar', 'foo', 'bar'))
+			(new Sql\Join('foobar', 'foo', 'bar'))
 				->right()
 				->sql()
 		);
 	}
 }
 
-(new BuiltJoin)->run();
+(new Join)->run();
