+++ Tests/Unit/Limit.phpt
@@ -11,21 +11,21 @@ require __DIR__ . '/../bootstrap.php';
 /**
  * @testCase
  */
-final class PreparedLimit extends Tester\TestCase {
+final class Limit extends Tester\TestCase {
 
 	public function testReturningEmptyStringWithMaximumInteger(): void {
 		Tester\Assert::same(
 			'',
-			(new Sql\PreparedLimit(PHP_INT_MAX))->sql()
+			(new Sql\Limit(PHP_INT_MAX))->sql()
 		);
 	}
 
 	public function testReturningLimitWithOffset(): void {
 		Tester\Assert::same(
 			'LIMIT ? OFFSET ?',
-			(new Sql\PreparedLimit(10))->sql()
+			(new Sql\Limit(10))->sql()
 		);
 	}
 }
 
-(new PreparedLimit)->run();
+(new Limit)->run();
