+++ Tests/Unit/OrderBy.phpt
@@ -11,35 +11,35 @@ require __DIR__ . '/../bootstrap.php';
 /**
  * @testCase
  */
-final class BuiltOrderBy extends Tester\TestCase {
+final class OrderBy extends Tester\TestCase {
 
 	public function testReturningOrderBySingleColumn(): void {
 		Tester\Assert::same(
 			'ORDER BY foo ASC',
-			(new Sql\BuiltOrderBy('foo'))->sql()
+			(new Sql\OrderBy('foo'))->sql()
 		);
 	}
 
 	public function testReturningOrderByManyColumns(): void {
 		Tester\Assert::same(
 			'ORDER BY foo, bar ASC',
-			(new Sql\BuiltOrderBy('foo', 'bar'))->sql()
+			(new Sql\OrderBy('foo', 'bar'))->sql()
 		);
 	}
 
 	public function testReturningAscendingOrder(): void {
 		Tester\Assert::same(
 			'ORDER BY foo, bar ASC',
-			(new Sql\BuiltOrderBy('foo', 'bar'))->ascending()->sql()
+			(new Sql\OrderBy('foo', 'bar'))->ascending()->sql()
 		);
 	}
 
 	public function testReturningDescendingOrder(): void {
 		Tester\Assert::same(
 			'ORDER BY foo DESC',
-			(new Sql\BuiltOrderBy('foo'))->descending()->sql()
+			(new Sql\OrderBy('foo'))->descending()->sql()
 		);
 	}
 }
 
-(new BuiltOrderBy)->run();
+(new OrderBy)->run();
