+++ Tests/Unit/Update.phpt
@@ -11,14 +11,14 @@ require __DIR__ . '/../bootstrap.php';
 /**
  * @testCase
  */
-final class PreparedUpdate extends Tester\TestCase {
+final class Update extends Tester\TestCase {
 
 	public function testReturningUpdateWithSingleColumn(): void {
 		Tester\Assert::same(
 			'UPDATE foo SET foo = :foo WHERE foo = :foo',
-			(new Sql\PreparedUpdate('foo'))
+			(new Sql\Update('foo'))
 				->set(['foo' => 'bar'])
-				->where(new Sql\BuiltWhere('foo = :foo'))
+				->where(new Sql\Where('foo = :foo'))
 				->sql()
 		);
 	}
@@ -26,9 +26,9 @@ final class PreparedUpdate extends Tester\TestCase {
 	public function testReturningOutputWithManyColumns(): void {
 		Tester\Assert::same(
 			'UPDATE foo SET foo = :foo, bar = :bar WHERE foo = :foo',
-			(new Sql\PreparedUpdate('foo'))
+			(new Sql\Update('foo'))
 				->set(['foo' => 'bar', 'bar' => 'foo'])
-				->where(new Sql\BuiltWhere('foo = :foo'))
+				->where(new Sql\Where('foo = :foo'))
 				->sql()
 		);
 	}
@@ -36,7 +36,7 @@ final class PreparedUpdate extends Tester\TestCase {
 	public function testReturningUpdateWithoutWhere(): void {
 		Tester\Assert::same(
 			'UPDATE foo SET foo = :foo',
-			(new Sql\PreparedUpdate('foo'))
+			(new Sql\Update('foo'))
 				->set(['foo' => 'bar'])
 				->sql()
 		);
@@ -46,7 +46,7 @@ final class PreparedUpdate extends Tester\TestCase {
 		Tester\Assert::same(
 			'UPDATE foo SET foo = :foo RETURNING foo, bar',
 			(new Sql\Returning(
-				(new Sql\PreparedUpdate('foo'))
+				(new Sql\Update('foo'))
 					->set(['foo' => 'bar']),
 				'foo',
 				'bar'
@@ -55,4 +55,4 @@ final class PreparedUpdate extends Tester\TestCase {
 	}
 }
 
-(new PreparedUpdate)->run();
+(new Update)->run();
